+++ src/skill-on-career/skill-on-career.service.ts
@@ -1,7 +1,9 @@
+import { UserRepository } from './../users/users.repository';
 import { CareerRepository } from './../career/career.repository';
 import { SkillRepository } from './../skill/skill.repository';
 import {
   BadRequestException,
+  ForbiddenException,
   Injectable,
   Logger,
   NotFoundException,
@@ -13,18 +15,20 @@ import {
   GetByCarrerIdDto,
 } from './dto';
 import { SkillOnCareerRepository } from './skill-on-career.repository';
+import { UserJwtPayload } from '../interfaces/jwt-payload';
 
 @Injectable()
 export class SkillOnCareerService {
   private logger: Logger = new Logger(SkillOnCareerService.name);
   skillOnCareerRepository: SkillOnCareerRepository;
-
+  private userRepository: UserRepository;
   private skillRepository: SkillRepository;
   private careerRepository: CareerRepository;
   constructor(private prisma: PrismaService) {
     this.careerRepository = new CareerRepository(this.prisma);
     this.skillRepository = new SkillRepository(this.prisma);
     this.skillOnCareerRepository = new SkillOnCareerRepository(this.prisma);
+    this.userRepository = new UserRepository(this.prisma);
   }
 
   async getByCareerId(dto: GetByCarrerIdDto) {
@@ -42,8 +46,15 @@ export class SkillOnCareerService {
     }
   }
 
-  async create(dto: CreateSkillOnCareerDto) {
+  async create(dto: CreateSkillOnCareerDto, user: UserJwtPayload) {
     try {
+      const userInfo = await this.userRepository.findById({
+        id: user.id,
+      });
+
+      if (!userInfo || userInfo.role !== 'ADMIN') {
+        throw new ForbiddenException('Access deny');
+      }
       if (dto.weight > 1 && dto.weight < 0) {
         throw new BadRequestException('Weight must be between 0 and 1');
       }
@@ -76,8 +87,15 @@ export class SkillOnCareerService {
     }
   }
 
-  async delete(dto: DeleteSkillOnCareerDto) {
+  async delete(dto: DeleteSkillOnCareerDto, user: UserJwtPayload) {
     try {
+      const userInfo = await this.userRepository.findById({
+        id: user.id,
+      });
+
+      if (!userInfo || userInfo.role !== 'ADMIN') {
+        throw new ForbiddenException('Access deny');
+      }
       return await this.skillOnCareerRepository.delete({
         where: {
           id: dto.id,
